Reduce memory usage of method tables
This is achieved by using `mrb_smap` as in instance variable tables. The
memory usage of `mrbtest` (full-core gembox) is as follows in the result
by Valgrind.

* Baseline: 131,437 allocs, 19,101,441 bytes allocated
* New:      131,419 allocs, 18,378,609 bytes allocated

I haven't measured the performance because I think it will be slightly
faster with the same tendency as instance variables. See the report for
instance variables below for details.

* https://shuujii.github.io/mruby-iv-benchmark